+++ CHANGELOG.md
@@ -44,10 +44,10 @@
 * Add `BTVenmoRequest`
 * PayPal
   * Fix memory leak in `BTPayPalDriver`
-  * Add `offerPayLater` to `BTPayPalRequest`
   * Add `BTPayPalCheckoutRequest`
   * Add `BTPayPalVaultRequest`
   * Add `tokenizePayPalAccount` method to `BTPayPalDriver`
+  * Add `offerPayLater` and `requestBillingAgreement` to `BTPayPalCheckoutRequest`
 
 ## 5.0.0-beta2 (2021-01-20)
 * Add SPM support for `BraintreeDataCollector` and `BraintreeThreeDSecure`